Clean VS2022 build steps and runtime settings

+++ src/ProxyDll/CharInfoFilter.cpp
@@ -123,16 +123,16 @@ int CCharInfoFilter::FilterRecvPacket(CPacket& pkt, CFilterContext& context)
 
 		if (pktMove.GetId() == m_wPlayerId)
 		{
-			m_bX = pktMove.GetX();
-			m_bY = pktMove.GetY();
+                        m_bX = static_cast<BYTE>(pktMove.GetX());
+                        m_bY = static_cast<BYTE>(pktMove.GetY());
 		}
 	}
 	else if (pkt == CWarpReplyPacket::Type())
 	{
 		CWarpReplyPacket& pkt2 = (CWarpReplyPacket&)pkt;
 
-		m_bX = pkt2.GetX();
-		m_bY = pkt2.GetY();
+                m_bX = static_cast<BYTE>(pkt2.GetX());
+                m_bY = static_cast<BYTE>(pkt2.GetY());
 
 		m_vPlayers.clear();
 		UpdateSuspendedFlag();
@@ -151,8 +151,8 @@ int CCharInfoFilter::FilterRecvPacket(CPacket& pkt, CFilterContext& context)
 			GetProxy()->send_packet(CCharacterSayPacket(m_szCharName, "--afk off"));
 		}
 
-		m_bX = xx;
-		m_bY = yy;
+                m_bX = static_cast<BYTE>(xx);
+                m_bY = static_cast<BYTE>(yy);
 
 		m_vPlayers.clear();
 		UpdateSuspendedFlag();
@@ -161,8 +161,8 @@ int CCharInfoFilter::FilterRecvPacket(CPacket& pkt, CFilterContext& context)
 	{
 		CCharStatsPacket& pkt2 = (CCharStatsPacket&)pkt;
 
-		m_bX = pkt2.GetX();
-		m_bY = pkt2.GetY();
+                m_bX = static_cast<BYTE>(pkt2.GetX());
+                m_bY = static_cast<BYTE>(pkt2.GetY());
 	}
 	else if (pkt == CUpdatePosSTCPacket::Type())
 	{
@@ -170,8 +170,8 @@ int CCharInfoFilter::FilterRecvPacket(CPacket& pkt, CFilterContext& context)
 
 		if (pkt2.GetId() == m_wPlayerId)
 		{
-			m_bX = pkt2.GetX();
-			m_bY = pkt2.GetY();
+                    m_bX = static_cast<BYTE>(pkt2.GetX());
+                    m_bY = static_cast<BYTE>(pkt2.GetY());
 		}
 	}
 	else if (pkt == CLevelUpPacket::Type())
@@ -339,8 +339,8 @@ int CCharInfoFilter::FilterSendPacket(CPacket& pkt, CFilterContext& context)
 	{
 		CUpdatePosCTSPacket& pkt2 = (CUpdatePosCTSPacket&)pkt;
 
-		m_bX = pkt2.GetX();
-		m_bY = pkt2.GetY();
+            m_bX = static_cast<BYTE>(pkt2.GetX());
+            m_bY = static_cast<BYTE>(pkt2.GetY());
 	}
 
 	return 0;
